Refactor: Organize code into app package
